Move test not using script to tests folder
